// Src/ShopperHolic.WebUI/ShopperHolic/src/app/models/stock/items/itemDetailed.ts
export class ItemDetailed{

    itemID: number = 0 ;
    itemCode: string = "";
    itemName: string = "";
    itemDescription: string = "";
    itemUnitPrice: number = 0;
    itemUnitPriceWithMaxDiscount: number = 0;
    itemAvailableQty: number = 0;
    itemReorderQtyReminder: number = 0;
    itemImageFilename: string = "";
    subGroupID: number = 0;
    subGroupCode: string = "";
    subGroupName: string = "";
    subGroupDescription: string = "";
    productGroupID: number = 0;
    productGroupCode: string = "";
    productGroupName: string = "";
    productGroupDescription: string = "";
    isFeaturedItem: boolean = false;
}